requirements.txt
# Core dependencies
streamlit>=1.28.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# PDF processing
PyMuPDF>=1.23.0

# Embeddings and ML
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4 
transformers>=4.30.0

# Vector search and text processing
langchain>=0.1.0
spacy>=3.6.0

# OpenAI
openai>=1.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# NLP models (download separately)
# python -m spacy download en_core_web_sm